refactor: P2-T2 - 从PracticeStepView提取Step3WindowGeometryHint并归档到Views/Practice/
